LayoutTests/imported/w3c/web-platform-tests/cookies/samesite/sandbox-iframe-subresource.https-expected.txt



FAIL SameSite cookies with intervening sandboxed iframe and subresources assert_equals: `samesite_lax=0.959107380167203` in request to `https://localhost:9443`. expected (undefined) undefined but got (string) "0.959107380167203"
